requirements.txt
Django
flake8
djangorestframeworkpygments